image-overlay/requirements.txt
image == 1.5.33
pillow == 9.5.0
functions-framework == 3.5.0
Flask ==3.0.3
google-cloud-storage ==2.16.0
google-auth == 2.29.0
gunicorn==20.1.0
python-dotenv ==1.0.1